modCommon.js -> modCommong.ts
Also re-wrote getDefaultMod10Weights to be a bit more clean and have a bit better performance.
